reticket-main/backend/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
black==22.1.0
click==8.0.3
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
factory-boy==3.2.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
pycodestyle==2.8.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2021.5
uritemplate==4.1.1
